Install nvim-lspconfig plugin
I've copied the add-plugin script over from my dotvim repo (my Vim 8
config).

I use it to install plugins as submodules, which I keep up to date with
the minpac package manager.

The packages.vim file contains minpac code that will keep the plugins
I've installed in this Neovim repository up to date. Note the `dir`
argument that's passed to minpac#init(). Without it minpac would update
the packages installed in the first directory of `packpath`, which is my
dotvim repo.

Those packages are updated when I run `:PackUpdate` in vim. I need
packages in this repository to be updated when I run `:PackUpdate` in
nvim.