Configure the gopls language server for Go
I'm following along with TJ and Bashbunni's Bash2Basics video series.
This config is from the first video, though I tweaked a couple of the
mappings slightly. [1]

I think `ctrl-]` is better than `gd` for jumping to definition; it's the
standard mapping for jump to definition. `gd` is traditionally used for
jumping to a local definition (i.e. in the current function).

However, if it turns out that LSP doesn't fully replace ctags I'll need
to use both, and will need a second mapping. `gd` would then make a lot
of sense.

[1] https://www.youtube.com/playlist?list=PLep05UYkc6wTWlugE_9Lj6JlLpvSBbkZ_