Setup completion with nvim-cmp
nvim-cmp is a lightweight plugin that's designed to work with other
plugins that provide completion "sources". e.g. cmp-nvim-lsp, cmp-buffer.

I've installed luasnips as nvim-cmp works with a snippet plugin, and TJ
thinks that out of the available options it's the way forward [1]. He's
also got a video dedicated to that (Take Tuesday episode 3).

The code in completion.lua was copied from [2].

[1] https://youtu.be/puWgHa7k3SY?t=3519
[2] https://github.com/hrsh7th/nvim-cmp#recommended-configuration