Define Ref command for finding references
I'd have gone for a keybinding, but I expect to change this as soon as
I've got going with Telescope. I imagine I'll replace it with a mapping
for `:Telescope lsp_references`, perhaps `<leader>tr` if I decide to
remove the command-t plugin. I could have a bunch of Telescope shortcuts
in the `<leader>t` namespace.