Share status line between all windows
This option just landed in the last nightly Neovim release. You can also
control the way the splits between windows are rendered with:

    :highlight WinSeparator guibg=None

I haven't needed to do that with my base16 gruvbox-dark-pale theme;
there's no visible background on the separator anyway. Setting it to
None just made the line brighter, which wasn't a good look.

See TJ's video for a demo:

https://www.youtube.com/watch?v=jH5PNvJIa6o