Install Telescope fuzzy finder
The keymaps that I've defined here override the ones I've setup in my
dotvim repo for fzf.vim.

plenary.vim is a dependency; it's a shared library of reusable Lua for
building stuff in Neovim. telescope-fzf-native.nvim is optional, but
includes a re-implementation of fzf in C, so it runs fast.

Note that the telescope-fzf-native.nvim plugin needs `make` running when
first installed, and whenever it's updated.

minpac will take care of running `make` when it pulls down updates, but
whenever a new version is retrieved via `git submodule update` (this
will happen often as I've installed it as a submodule), minpac won't
pull down any updates and won't fire its post-update hook.

I'll sometimes find, therefore, that the plugin code is updated but that
the `.so` file hasn't been recompiled. Running `make` in the plugin
directory manually will serve as a workaround, but I need a better
solution.

See TJ's comments on why telescope-fzf-native.nvim is worth it:

https://youtu.be/guxLXcG1kzQ?t=1239