Make window border more visible
Since setting `laststatus=3` in 8c7b734 I've found it very difficult to
see the border between windows, in some colour themes. I've had a look
in the source for base16's woodland theme (for example) and WinSeparator
doesn't appear to be set. It would be nice if the border colours were
set as part of the theme though.