Redefine Telescope keybindings in Lua
I've been influenced here by Prime's argument that you don't want
consecutive keys in your mappings to be on the same hand, never on the
same finger.

I've gone for `<C-p>` for finding files because I'd quite like it on a
single keystroke. I had it on `<leader>ff` but I can see that giving me
RSI, long term. I can't use `<leader>f` because all my other fuzzy
finder mappings start with that. I'd go back to using `<leader>t` but I
like the fact that it's now in use by `vim-test`.

And `<C-p>` is going to work in a lot of other tools. Let's go with it.

I've updated the Vim 8 mappings (where I'm using fzf.vim) to be
equivalent.