Setup TJ's mapping to execute line of Vim config
See https://youtu.be/u6EKq6z0CRU?t=1509

Note that I've changed `getline(">")` to `getline(".")`, which is
documented as getting the text of the current line. I couldn't find any
docs on what `getline(">")` returns, and it didn't work for me when
testing it.

It's not working properly when I highlight multiple lines of Vimscript
either, it just runs the first line in the visual block. I haven't tested
multiple lines of Lua yet.