Add keymaps for fuzzy finding of commands and help
I've moved the file browser to <leader>B because I think I'll find
reviewing commits on the current buffer really useful, and I want to
keep it. I could have moved it to <leader>fC, but I really like that for
searching the commands (especially as all non-builtin commands need to
start with a capital letter).