Configure Python provider to use system Python
Without this setting the Neovim Python provider was finding my
asdf-installed versions of Python, and virtual envs in `./.venv`.
Neither of them have the pynvim module installed.

In my infrastructure repo I'm installing the pynvim Python module with
the system version of Pip, which is why I'm using the system version of
Python. If appropriate it might make sense to switch this over to the
asdf install, but as I currently use Python in Neovim (for the `:Black`
command) there's no need to keep it that up to date.

Use the `:checkhealth` command for diagnostics, if you think this is no
longer setup properly.