Neoformat prefers prettier in node_modules
Neoformat is reported to ship with a prettier binary, though when I
tested it with `:Neoformat prettier` it returned an error as the command
wasn't found.

No problem, I only want to be running Prettier in projects where
Prettier is installed and configured.

This setting should ensure that if Prettier is installed in a project,
Neoformat will run that version.

To install prettier, run one of these commands:

    npm install -D prettier
    yarn add -D prettier