Group plugins by role like lsp-zero.nvim
The setup docs for lsp-zero.nvim include a snippet of code that
specifies the LSP, Autocompletion and Snippets plugins in groups. Not a
bad idea in itself, and it makes it easy for me to compare my config to
that of lsp-zero.

As the author of lsp-zero points out, if you've already got all the same
stuff installed, you don't need it.