Set base16 colorscheme from external config file
These Lua functions have replaced this Vimscript, from my dotvim
repository:

    " base16-theme and vim-airline-themes
    let base16colorspace=256
    let theme = 'base16-' . readfile(expand('~/.config/base16-theme'))[0]
    execute('colorscheme ' . theme)
    let g:airline_theme = join(split(theme, '-'), '_')

It feels nicer in Lua, and it's good that the code runs in callbacks
after the relevant plugins are loaded.

I've moved the highlight command that sets the WinSeparator colour in
here too, as it only makes any sense in the context of the base16 themes
that I've been using.