Load Telescope on demand with lazy.nvim
When specifying dependencies the lazy.nvim docs say you can either
specify a plugin spec (i.e. a Lua table) or the name of a plugin that's
defined elsewhere.

They don't go into details on what "name" means in that context. I've
experimented, and have found you can use the following in a plugin's
dependencies table:

- the "user/repo" string that identifies the repository

- the "repo" part of "user/repo"

- a name that you've made up, and specified elsewhere in the plugin's
  spec with the `name` property.

So I'm going with "user/repo", as that's how you'd specify it if there
was no config required for that plugin. It seems less likely to need to
change under refactoring.