Install Prime's Harpoon plugin
I've setup keymaps to allow me to jump to files by number with `<leader>n`
or to cycle between them with `<Alt-]>` and `<Alt-[>` (next and previous).

Show and hide the Harpoon list with `<A-p>.`

Add a file to the project's Harpoon list with `<C-A-p>.` Remove a file
by opening the list window and deleting it (it's a normal buffer).