Disable key ordering check in yamlls
The default setting for this new-ish check was to enable it when it
first released. Since then it's been disabled by default, but a new
package hasn't been released yet. So I'll need this custom for yamlls
for a short while.

The current version is 1.12.0, and any release later than that should
allow me to revert to default config.

Use `:Mason` to check for updates.