Replace LSP and nvim-cmp setup with lsp-zero
I was initially inspired to try this by a video from Prime [1], but the
material I've based my config on is the lsp-zero docs ([2], [3], and [4]).
That's because lsp-zero 2.x has been released since Prime made his video,
and the API has improved. It now seems simpler, and exposes the configuration
APIs of the underlying plugins. Nice!

[1] https://www.youtube.com/watch?v=w7i4amO_zaE
[2] https://github.com/VonHeikemen/lsp-zero.nvim
[3] https://github.com/VonHeikemen/lsp-zero.nvim/blob/v2.x/doc/md/lsp.md
[4] https://github.com/VonHeikemen/lsp-zero.nvim/blob/v2.x/doc/md/autocomplete.md

I've also remapped the Emmet leader key, which defaults to `<C-y>` (the
same mapping that lsp-zero sets up for choosing the selected text). The
conflict had the irritating effect of inserting "Y" at the position of
the cursor for a split second, before replacing it with the completed
text, as Vim waited to see if I was about to enter a full Emmet mapping.

For Emmet's new leader I've gone with `<A-e>`.