Add autocmd to run checktime when focusing terminal
The checktime command works in tandem with the `autoread` setting (which
is set in `config/options.lua`) to ensure that if a file has changed on
disk since the buffere was modified, that its buffer automatically
reloads.

There's much discussion about which events to bind this to, and a few
alternative approaches. If these two events don't turn out to be
frequent enough, this is a good place to look for alternatives:

https://unix.stackexchange.com/questions/149209/refresh-changed-content-of-file-opened-in-vim/383044#383044