Add mapping for Telescope grep_string
The custom function that uses input() to prompt for a string, then
launches "Telescope grep_string" (that I took from one of Prime's
videos) seems useful, but I might not use it as much as :Rg.

The builtin approach to grep_string, which runs it on the word under the
cursor, feels like it might be more useful to me. So I've added that,
and tweaked the keymaps to reflect how useful I expect to find them.