Re-enable cmp-nvim-lua
I had this enabled before switching to lsp-zero. Note that I've
specified a total of 5 sources in this commit. It's the four defaults
that come with the "recommended" source, plus nvim_lua.

The recommended source appears to configure a bunch of other settings on
top of the list of sources, so I'm leaving it in.