Move Neoformat autocmd next to plugin config
Note that I've not defined the autocommand inside Neoformat's config
handler. That's because the plugin only gets loaded (and the config
therefore only runs) after the `:Neoformat` command has run. So it'd be
a catch 22, and the autocmd wouldn't get run automatically.

It feels cleaner to put plugin specific autocommands close to the code
that loads the plugin though.