Disable signcolumn
I don't think I enabled this setting because I was using any plugins
that put signs in the gutter; I suspect it was just because I liked the
look of some white space to the left of the line numbers.

With this turned off, I can get 3 x 80 col buffers in a maximised
terminal on my desktop machine, at a good font size. And that feels
worth trying.