Change keymap for signature_help to gh
I think I'm more likely to remember that than gs, where is was for
"signature". And you don't type both characters with the same hand, so
that's a win.

With pyright on the standard library, I'm finding signature_help does
the same thing as pressing K (hover). But maybe it's a bit different
with other language servers?