Disable hard wrapping mail mode
When writing emails I've always enjoyed (preferred) hard wrapping, as
that is a traditional default. Modern mail clients don't do it though,
and in order to gain some more consistency with emails that I write in
the Fastmail user interface, I'm disabling hard wrapping for mails
written in Vim too.