Remove ts_ls from lsp.setup_servers call
I'm not sure that it should ever have been there in the first place. I'm
getting a warning whenever I start nvim with this line in place, and I'm
finding that if I edit a .ts file that the language server seems to work
fine. So let's drop it.

Bug fixes are discussed online, referencing these links:

- https://github.com/neovim/nvim-lspconfig/pull/3232#issuecomment-2331025714
- https://github.com/williamboman/mason.nvim/issues/1784#issuecomment-2336761735
- https://www.reddit.com/r/neovim/comments/1f9iakw/lspconfig_renamed_tsserver_to_ts_ls_what_to_do_to/

Haven't read all that, I'm not doing TypeScript right now, and I want to
squash this frigging warning.