Change leader key to space
I've been meaning to try this for years, after Prime pointed out
the problem with using a comma.

The problem with it is that it makes navigation with f/F/t/T a bit
frustrating. You press ; to repeat the previous jump and , to backtrack.
If you're also using , for leader then Vim waits to see whether you're
about to type more keys before it backtracks.