Remove vim-ruby
The [[, ]], [m, and ]m mappings are really useful, but they're not
working in Neovim 0.10.2 for some reason. I get this error when trying
to use them:

    E117: Unknown function: <SNR>77_searchsyn

The mappings contain the text `<SID>`, whose help sheds some light on
what's going on:

> The string "<SID>" can be used in a mapping or menu.  This is useful if you
> have a script-local function that you want to call from a mapping in the same
> script.
>    When executing the map command, Vim will replace "<SID>" with the special
> key code <SNR>, followed by a number that's unique for the script, and an
> underscore.  Example:
>         :map <SID>Add
> would define a mapping "<SNR>23_Add".

From that we can infer that the `s:searchsyn` function isn't found,
despite being defined in the vim-ruby plugin. Not sure what's going on,
but this stuff clearly used to work fine, and it's not been touched in
over a decade.

With this plugin removed, I lose support for [[, ]], etc. They don't
appear to be supported in Neovim with plugin support.