Remove nvim-treesitter-context
I'm finding it that the way it formats Ruby code is really quite janky,
and the colours aren't different enough from the default background
(regardless of which theme I've tried) for it to be obvious that it's
decided to insert a bunch of bullcrap (generally highly verbose
multi-line comments) into more than half the screen.

Not worth the (very minor) benefit of knowing what you're currently
inside. Particularly in Ruby, where some dick has probably written a
module that they're going to start blatting into some other object
anyway, and this plugin doesn't help you work that out.