Configure Telescope layout for input at top
I liked this aspect of the "dropdown" theme that I was using when
finding files, but I hadn't configured it everywhere, and I wasn't
enamoured with dropdown's fairly short list of files to search through.
Also, the preview pane was quite small.

This approach sets defaults across all Telescope pickers.

Note the flex strategy too; it's like "horizontal" on wide windows (i.e.
with a preview) but adjusts behaviour to suit window width.