Increase width of Telescope's horizontal preview
The preview window was quite narrow, particularly when using it to
quickly browse help topics.

preview_width sets the fraction of the pane that's used for the preview;
making it wider than 50% works nicely, but it looks weird.

So I've increased the width of the entire pane, which gives everything
more space.