Install the gitsigns plugin
I've had to re-enable the sign column to get this working, which will
add some useful whitespace to the left of my line numbers (they do but
up against the edge of the terminal).

I note that when I disabled the signcolumn I was enjoying having 3
columns of 80 characters of text on my desktop monitor, but I don't
really use that feature much. Also, when doing client work (which is
what I should mostly be using), clients don't tend to have good coding
styles so I can't use 3 columns anyway.

I should focus on my workflow with just one buffer visible (next to a
browser), or when deep in the TDD flow, two side-by-side buffers.