Stop loading fugitive lazily
I had it starting when any of the listed commands were run. That's just
a fraction of what Fugitive can do, so it's a daft way to do it. With
this new approach Vim will autocomplete any :G* commands that I start
typing immediately, which is good for discovery of Fugitive's features.