Choose when Telescope's file preview goes vertical
The vertical layout's value of preview_cutoff is used as the default
value for flip_lines if it's not set.

I set preview_cutoff instead of flip_lines because when I removed
preview_cutoff in favour of flip_lines, I found that the vertical
preview was hidden much sooner than I'd like (as the window's height
reduced).