Hide the cmdline when not in use
I'll have to see how I get on with this. I like the extra row of real
estate, but I notice that :commands now get typed over the top of the
status line. Do I ever find the filename (shown in the status line)
useful when crafting a command? Not sure. And I'll often be able to just
type % if I do...