Install cloak.nvim to protect secrets
It hides variable values if you accidentally open an env file when it's
visible to people who shouldn't be able to see it.

Use :CloakPreviewLine to show the value of a single line. Other :Cloak*
commands are available.

Note that we're passing an empty `opts` table; that'll trigger lazy to
run the plugin's setup() for us. Without that cloak doesn't do anything.