Convert Telescope keymap config native Vim API
The Lazy keys property is nice (it can autoload plugins when you press a
keystroke), but I feel there's a lot to be said for using the builtin
API.

Firstly, you become familiar with it.

Secondly, you can add a keymap and try it out without restarting Vim if
you use :.lua to source the line (for which I also have a keymap!).