Convert Telescope keymaps to functions
Commands work great, but calling the functions directly feels a little
cleaner.

I then had to document what all the functions did, which made the code a
lot more verbose, but the benefit of this is that I get to choose how to
describe each keymap. That's very useful in which-key's overlay.