Define custom live grep Telescope behaviour
This is based on the code that TJ demonstrated in his 2024 Advent of
Code [1].

He limits the output from ripgrep using its -g option to pass a glob
pattern. I usually prefer using -w and passing a path to search through,
and I'll modify it to do that in another commit.

[1] https://www.youtube.com/watch?v=xdXE1tOT-qg&t=191